demo/python/requirements.txt
# python==3.12.3
pygame==2.6.0
tqdm==4.66.4
numpy==2.0.0
pillow==10.4.0
# dependencies for GPLT
matplotlib==3.9.1
rich==13.7.1
PyQt5==5.15.10
